Allow T mainline connectivity code to link libnativehelper.
All the connectivity code uses libnativehelper to register JNI
methods. In T more connectivity code is moving out of the
framework. Allow that code to use libnativehelper_compat_libc++.

Allow everything in packages/modules/Connectivity, similar to
what is allowed for ExtServices and NetworkStack.
Also include temporary locations in f/b/packages/ConnectivityT,
which is where the code is at the moment.

Test: m
Change-Id: I582feaaa51f9eeae7dd26193865e7644422bf613