Move NIOAccess methods directly into JNI for the JVM
RNG and LayoutLib have a different NIOAccess Java delegate classes. Move
the logic into native code to avoid having to delegate to a Java class.
JNI call overhead is not an issue because none of these methods are in
the hot path. They are just used in the setup logic of
Bitmap.copyPixels{To,From}Buffer.

Test: ShadowNativeBitmapTest in Google3
Bug: 311779352

Change-Id: I42cbffc071de189e6609937f6116f4ad3f30ece4